Refactoring of pages
